Update Full-Site-Page-BP-example.html
To avoid confusion on status of what is spec'd and what is not, Subbu removed three attributes from the current TP BP example that did not catch the schema 3.3 release. We will add these back when 3.4+ catches them. 

- ownershipFundinGrants
- diversityStaffingReport
- refLocalNationalRequirements

Note: foundingDate is already defined as property of Organization and hence usable as markup for NewsMediaOrganization, even we have not formally specified as Trust Markup yet for the BP Founding Date protocol attribute, along with the 3.3 release.