Update and rename Snippet-Citations_References.html to Basic_Snippet-Citations_References.html
Updated the citations markup from an array of links, with to array of linked marked up with  "CreativeWork". This is now valid for MVP  and is a basic one for newsrooms that don't have a CMS-UX workflow to categorize the citations further internally as documents, datasets, secondary news etc. A separate example will cover the case for newsrooms with that capacity.