Update Snippet-Author-Info.html
Removing worksFor after feedback from newsrooms. (The @type NewsMediaOrganization where the org is named as publisher and the author's declaration by name and jobTitle should be good enough to establish that a relationships exists as needed to be disclosed/signalled by attribute #9 of the TTP Author Producer/info -- Affiliation with Publisher/Job Title.