Update NYT-POW-opinion-no-author.html
Add the basic Organization-is-author schema to this piece. Economist pieces are doing this now, since all their pieces don't have author names. The fuller no-byline, brand-author exception author trust indicator schema is still pending discussion at schema.org and we can update this example after that is ready. For now this is ready to send for testing after